# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event-driven Notification API",
        "description": "API for creating, tracking, and validating notifications.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/providers/{provider}/callbacks": {
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Handle provider delivery callbacks",
                "operationId": "16f9e86fa4f2c44dc473c69b8aa52f45",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Callback accepted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Callback verification failed."
                    },
                    "404": {
                        "description": "Provider or notification not found."
                    },
                    "422": {
                        "description": "Unsupported receipt status."
                    }
                }
            }
        },
        "/api/templates": {
            "get": {
                "tags": [
                    "Templates"
                ],
                "summary": "List templates",
                "operationId": "e7bff3994948a3b495a185984da5e485",
                "parameters": [
                    {
                        "name": "channel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sms",
                                "email",
                                "push"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paged templates list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Template"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Templates"
                ],
                "summary": "Create a template",
                "operationId": "b7aa5396b7d49d88a8b9a357701f3cee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTemplateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/templates/validate": {
            "post": {
                "tags": [
                    "Templates"
                ],
                "summary": "Validate and render a template",
                "operationId": "76c2c3be9f196d510e6a8c85858e6dc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateValidationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template rendered.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rendered": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
        "/api/templates/{template}": {
            "get": {
                "tags": [
                    "Templates"
                ],
                "summary": "Get template by id",
                "operationId": "ed6dce44f72216db9c0ab0a8daa9e37d",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Template not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Templates"
                ],
                "summary": "Delete a template",
                "operationId": "cac7818682de619f7c6aa7dcdf74ed74",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Template deleted."
                    },
                    "404": {
                        "description": "Template not found."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Templates"
                ],
                "summary": "Update a template",
                "operationId": "f1d216346a9951c55b6b280f8bab682e",
                "parameters": [
                    {
                        "name": "template",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTemplateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreTemplateRequest": {
                "required": [
                    "name",
                    "channel",
                    "body"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email",
                            "push"
                        ]
                    },
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TemplateValidationRequest": {
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "sample_variables": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
            "UpdateTemplateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email",
                            "push"
                        ]
                    },
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Template": {
                "required": [
                    "id",
                    "name",
                    "channel",
                    "body",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "sms",
                            "email",
                            "push"
                        ]
                    },
                    "body": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}